% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{cluster_prep}
\alias{cluster_prep}
\title{Identify the originating package of an S3 or S4 model object
and load it on a PSOCK cluster}
\usage{
cluster_prep(model, cl)
}
\arguments{
\item{model}{An R model object (S3 or S4)}

\item{cl}{A PSOCK cluster (created with parallel::makeCluster)}
}
\value{
Invisibly returns the package name loaded
}
\description{
Identify the originating package of an S3 or S4 model object
and load it on a PSOCK cluster
}
\details{
For internal use
}
\keyword{internal}
