% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{extract_namespace}
\alias{extract_namespace}
\title{Extract the Namespace from a Model Call}
\usage{
extract_namespace(x)
}
\arguments{
\item{x}{A fitted model object.}
}
\value{
A character string with the namespace (package name), or `NULL`
  if the namespace cannot be determined.
}
\description{
This function attempts to extract the package namespace from the call
used to fit a model object, assuming the function was called using the
`pkg::fun()` syntax.
}
\details{
For internal use
}
\keyword{internal}
