% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\name{hab}
\alias{hab}
\title{Example raster}
\format{
A binary SpatRaster object
\describe{
  \item{hab}{ --> A binary raster}
}
}
\description{
Example habitat raster for optimizing scales of effect
}
\examples{
hab <- terra::rast(system.file("extdata",
                   "hab.tif", package = 'multiScaleR'))
}
