% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\name{landscape}
\alias{landscape}
\title{Simulated raster}
\format{
`landscape_rast`

A spatRaster object with three surfaces:
\describe{
  \item{land1}{ --> A binary landscape surface with low autocorrelation}
  \item{land2}{ --> A continuous landscape surface with low autocorrelation}
  \item{land3}{ --> A continuous landscape surface with high autocorrelation}
}
}
\description{
Raster data for use with vignette example
}
\examples{
land_rast <- terra::rast(system.file("extdata",
                         "landscape.tif", package = 'multiScaleR'))

}
