% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiScaleR-package.R
\docType{package}
\name{multiScaleR-package}
\alias{multiScaleR}
\alias{multiScaleR-package}
\title{multiScaleR}
\description{
This package is for optimizing scales of effect when modeling ecological processes in space. Specifically, the scale parameter of a distance-weighted kernel distribution is identified for all environmental layers included in the model.
}
\details{
\if{html}{
  \figure{logo.png}{options: width=150 alt='Package logo'}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wpeterman/multiScaleR}
  \item Report bugs at \url{https://github.com/wpeterman/multiScaleR/issues}
}

}
\author{
\strong{Maintainer}: Bill Peterman \email{Peterman.73@osu.edu} (\href{https://orcid.org/0000-0001-5229-9268}{ORCID})

}
