% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_method.R
\name{plot.multiScaleR}
\alias{plot.multiScaleR}
\title{Plot method for multiScaleR objects}
\usage{
\method{plot}{multiScaleR}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multiScaleR}.}

\item{...}{Arguments to modify the plot. See Details.}
}
\value{
A list of ggplot2 objects.
}
\description{
Plot kernel weight distributions from optimized \code{multiScaleR} objects.
}
\details{
Supported arguments include:
\itemize{
  \item \code{prob}: Cumulative weight cutoff for distance scale (default = 0.9).
  \item \code{scale_dist}: Logical; add vertical line for distance scale (default = TRUE).
  \item \code{add_label}: Logical; annotate scale distance and CI (default = TRUE).
}
}
\examples{
\dontrun{
plot(x)

plot(x, prob = 0.95)

plot(x, scale_dist = FALSE)

plot(x, scale_dist = TRUE, add_label = FALSE)
}

}
\seealso{
\code{\link[multiScaleR]{plot_kernel}}
}
