% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.multiScaleR_data}
\alias{print.multiScaleR_data}
\title{Print method for multiScaleR_data}
\usage{
\method{print}{multiScaleR_data}(x, ...)
}
\arguments{
\item{x}{A \code{multiScaleR_data} object}

\item{...}{Ignored}
}
\description{
Print method for objects of class \code{multiScaleR_data}.
}
