% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.summary_multiScaleR}
\alias{print.summary_multiScaleR}
\title{Print method for summary_multiScaleR}
\usage{
\method{print}{summary_multiScaleR}(x, ...)
}
\arguments{
\item{x}{A \code{summary_multiScaleR} object}

\item{...}{Ignored}
}
\description{
Print method for objects of class \code{summary_multiScaleR}.
}
