% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{summary.multiScaleR}
\alias{summary.multiScaleR}
\title{Summarize multiScaleR objects}
\usage{
\method{summary}{multiScaleR}(object, ...)
}
\arguments{
\item{object}{An object of class \code{multiScaleR}.}

\item{...}{Optional arguments passed to the method (e.g., \code{prob} for cumulative kernel weight threshold).}
}
\value{
An object of class \code{summary_multiScaleR}.
}
\description{
Summarizes output from \code{multiScale_optim}.
}
