% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\name{surv_pts}
\alias{surv_pts}
\title{Spatial sample points}
\format{
An sf class point object:
\describe{
  \item{pts}{ --> 100 spatial point locations}
}
}
\usage{
data(surv_pts)
}
\description{
Example point file for use with vignette document example
}
