% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft_smooth.R
\name{fft_convolution}
\alias{fft_convolution}
\title{Fast Fourier Transformation}
\usage{
fft_convolution(x, kernel, fun = "mean", na.rm = TRUE)
}
\arguments{
\item{x}{A matrix of the raster layer}

\item{kernel}{A weight matrix for the smoothing}

\item{fun}{Only calculates mean}

\item{na.rm}{Always remove NA values from mean calculation}
}
\value{
A matrix
}
\description{
Kernel smoothing with fft
}
\keyword{internal}
