% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.multiScaleR}
\alias{print.multiScaleR}
\title{Print method for multiScaleR}
\usage{
\method{print}{multiScaleR}(x, ...)
}
\arguments{
\item{x}{A \code{multiScaleR} object}

\item{...}{Ignored}
}
\description{
Print method for objects of class \code{multiScaleR}.
}
